customer-api/customer/requirements.txt
requests
flask-lambda-python36
flask
Werkzeug==2.0.3
pandas
boto3